Add files via upload
Created a VPC that contained a private and public subnet, IGW, and NAT. Configured Route Tables associated with a public subnet for internet-bound traffic to be directed to the IGW for direct internet access. Configured Route Tables associated with a private subnet for isolated resources to securely connect to the internet through a NAT Gateway. Launched a Bastion Host in a public subnet for resource-based secured access to the private subnet. Evaluated the operation of the private subnet’s ability to communicate with the internet.